just-another-voice-agent-livekit/pyproject.toml
[project]
name = "just-another-voice-agent-livekit"
version = "0.1.0"
description = ""
authors = [
    {name = "Priyanshu Yashwant Deshmukh",email = "priyanshudeshmukh4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector] (>=1.0,<2.0)",
    "elevenlabs (>=2.5.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "just_another_voice_agent_livekit", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
